Rename ekf model files
